Extract directory link traversal partial
